backend/requirements.txt
fastapi
fastapi-cache2==0.2.1
apscheduler==3.10.4
redis==5.0.1
jinja2==3.1.4
uvicorn[standard]
httpx[http2]
asyncpg
sqlalchemy[asyncio]
alembic
python-dotenv==1.0.1
uvicorn==0.27.1
asyncpg==0.29.0
python-multipart==0.0.9
pdfminer.six==20231228
anthropic
PyGithub
pydantic
python-jose[cryptography]
prometheus-client
